migrated sybil-gate-template pallet tests to construct_runtime
