+++ sybil-gate-template/src/lib.rs
@@ -204,4 +204,6 @@ decl_error! {
 }
 
 #[cfg(test)]
+mod mock;
+#[cfg(test)]
 mod tests;
