+++ sybil-gate-template/src/tests.rs
@@ -15,52 +15,14 @@
 // along with Encointer.  If not, see <http://www.gnu.org/licenses/>.
 
 use super::*;
-use crate::{Config, Module};
+use crate::mock::{TestRuntime, new_test_ext, Origin};
 use frame_support::assert_ok;
 use sp_core::H256;
-use sp_runtime::{
-    testing::Header,
-    traits::{BlakeTwo256, IdentityLookup},
-};
 use xcm_executor::traits::Convert;
 
 use test_utils::*;
-use frame_support::dispatch::DispatchInfo;
 
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub struct TestRuntime;
-
-pub type System = frame_system::Pallet<TestRuntime>;
-
-impl_frame_system!(TestRuntime);
-impl_balances!(TestRuntime, System);
-impl_outer_origin_for_runtime!(TestRuntime);
-
-impl Config for TestRuntime {
-    type Event = ();
-    type Call = EmptyCall;
-    type XcmSender = ();
-    type Currency = balances::Pallet<TestRuntime>;
-    type Public = <Signature as Verify>::Signer;
-    type Signature = Signature;
-}
-
-#[derive(Clone, PartialEq, Eq, Debug, Decode, Encode)]
-pub struct EmptyCall(());
-impl GetDispatchInfo for EmptyCall {
-    fn get_dispatch_info(&self) -> DispatchInfo {
-        Default::default()
-    }
-}
-
-fn new_test_ext() -> sp_io::TestExternalities {
-    let storage = frame_system::GenesisConfig::default()
-        .build_storage::<TestRuntime>()
-        .unwrap();
-    storage.into()
-}
-
-type SybilGate = Module<TestRuntime>;
+type SybilGate = crate::Module<TestRuntime>;
 
 #[test]
 fn faucet_works() {
